requirements.txt
streamlit>=1.28.0
psycopg2-binary>=2.9.9
spacy>=3.7.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
pytest>=7.4.0
